Resolve built-in schemas at package boundary
