+++ crates/ironclaw_host_runtime/tests/tool_surface_contract.rs
@@ -399,10 +399,19 @@ async fn visible_surface_uses_caller_provider_trust_not_host_trust_policy() {
 
 #[tokio::test]
 async fn visible_surface_resolves_builtin_first_party_input_schema_refs() {
+    let package = builtin_first_party_package().unwrap();
+    assert_eq!(package.capabilities.len(), 14);
+    for descriptor in &package.capabilities {
+        assert!(
+            descriptor.parameters_schema.get("$ref").is_none(),
+            "{} should be registered with a concrete input schema, got {:?}",
+            descriptor.id,
+            descriptor.parameters_schema
+        );
+    }
+
     let mut registry = ExtensionRegistry::new();
-    registry
-        .insert(builtin_first_party_package().unwrap())
-        .unwrap();
+    registry.insert(package).unwrap();
     let runtime = runtime_with(registry, Arc::new(GrantAuthorizer));
     let builtin_effects = vec![
         EffectKind::DispatchCapability,
